# Fix: optimize the way we do our calls to the API 