--- a/lib/ontologies_api_client/http.rb
+++ b/lib/ontologies_api_client/http.rb
@@ -61,7 +61,7 @@ module LinkedData
         invalidate_cache = params.delete(:invalidate_cache) || false
 
         begin
-          puts "Getting: #{path} with #{params}" if $DEBUG
+          puts "Getting: #{path} with #{params}" if $DEBUG_API_CLIENT
           begin
             response = conn.get do |req|
               req.url path
@@ -87,7 +87,7 @@ module LinkedData
             obj = recursive_struct(load_json(response.body))
           end
         rescue StandardError => e
-          puts "Problem getting #{path}" if $DEBUG
+          puts "Problem getting #{path}" if $DEBUG_API_CLIENT
           raise e
         end
         obj
@@ -143,7 +143,7 @@ module LinkedData
       end
 
       def self.delete(id)
-        puts "Deleting #{id}" if $DEBUG
+        puts "Deleting #{id}" if $DEBUG_API_CLIENT
         response = conn.delete id
         raise StandardError, response.body if response.status >= 500