--- a/lib/ontologies_api_client/read_write.rb
+++ b/lib/ontologies_api_client/read_write.rb
@@ -90,7 +90,7 @@ module LinkedData
         HTTP.get(self.id, invalidate_cache: true) if self.id
         session = Thread.current[:session]
         session[:last_updated] = Time.now.to_f if session
-        refresh_cache
+        # refresh_cache
       end
 
       def refresh_cache